refactor(clusters): address full code review (H1-H3 + M1-M8)
High-priority fixes:
- H1 case-insensitive env grouping: validator now lower-cases the env
  value (and trims whitespace) so "PROD" / "prod" / "  PROD  " all
  collapse into the same group. groupByEnv also normalizes its key.
- H2 K8s auto-connect explicit labels: routers/k8s_clusters.py now
  passes labels={env: default} explicitly when creating the auto-link
  ConnectionProfile, so the contract no longer relies on the model
  validator default.
- H3 EditConnectionDialog useEffect dep narrowed to connection?.id so
  refetched profile objects don't clobber in-flight edits.

Componentization (M2 + M5 + M3 + M4):
- Split (main)/clusters/page.tsx (was 600+ lines, now ~110) into:
    components/clusters/{
      labels.ts            # ENV_LABEL_KEY, DEFAULT_ENV_VALUE,
                           # LabelEntry, labelsToEntries, entriesToLabels
      envSort.ts           # ENV_PRIORITY, envSortKey, compareEnv
      clusterRows.ts       # ClusterRow type, mergeRows, groupByEnv
      PhaseBadge.tsx       # phaseTone() + <PhaseDot/>
      NameLink.tsx
      EnvSectionHeader.tsx
      EnvSection.tsx
      ClusterTable.tsx
      ClusterCard.tsx
      ViewToggle.tsx
      ClustersSkeleton.tsx
      ClustersEmptyState.tsx
    }
  page.tsx now contains hooks + groupByEnv + dialog wiring only.
- Extracted hexToRgba into lib/utils.ts (M3 + M8). Returns null on
  malformed input so callers can branch instead of silently producing
  invalid CSS.
- Pulled label-domain primitives into components/clusters/labels.ts so
  LabelsCell no longer imports from LabelsEditor (display vs editor
  decoupled).
- Add / Edit Connection dialogs share useConnectionForm() hook +
  <ConnectionFormFields/> component (~80 lines of duplication removed).
  Edit dialog disables credential fields via showCredentials=false.

API hygiene:
- M1: removed duplicate _validate_labels validator from
  ConnectionProfileResponse — only construction path (from_profile)
  receives an already-normalized profile, so the duplicate was dead
  code that misled readers.
- M7: replaced over-generic _decode_json_column with a narrower
  _decode_json_dict (returns dict[str, Any] always). Keeps support for
  asyncpg pre-parsed JSONB and stays type-narrow for the labels call
  site.
- M6: hide the Labels cell entirely for unlinked ACKO rows in both
  table and card (previously rendered an editable-looking env=default
  chip the user couldn't actually edit).

Tests:
- env value lower-case normalization regression test.
- SQLite migration idempotency test (init_db twice over the same file).
- Whitespace-only label keys are dropped.
